requirements.txt
fastapi==0.116.1
httpx==0.28.1
matplotlib==3.10.1
mlflow==2.20.1
numpy==2.2.4
onnx==1.17.0
onnxruntime==1.20.1
pandas==2.2.3
pillow==11.1.0
prometheus-client==0.21.1
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
ruff==0.11.0
scikit-learn==1.6.1
torch==2.6.0
torchvision==0.21.0
uvicorn[standard]==0.34.0
pytest==8.3.5